Labs 1,2,3 Completed
Lab 1: Changed second column to * instead of Auto, so it takes up all available space, then changed the TextBox HorizontalAlignment to Stretch so they resize when the window does.

Lab 2: StackPanel only lab, used nested StackPanels. Outermost is vertical, which has the title Lorem Ipsum, then it has a horizontal stack panel, which has a left for the text and right for the rectangles.

Lab 3: 3x2 Grid, one row for each title and text/rectangle display. One column for the titles and one for the text/rectangles. The text/rectangle displays are vertical StackPanels with text above a horizontal StackPanel of rectangles.