Lab 4 Completed
Used a Grid outer container, with 2 rows and 1 column. Top row has a relative panel with the hamburger button, back button, and Financial/Food text block aligned to the left, while the search button and text box are aligned to the right.

The bottom row contains a Split View, where the pane has a list box of 2 items: financial and food. When clicked, these navigate to their respective pages. The content has a frame which displays the page of the selected choice in the pane, either financial or food.