Lab5 Finished
Added Image style for the logos so that they were all the same size and position.
Reformatted Schedule page to use the rows instead of having random stack panels.
Added logo and ticket  with text to complete page.
On Main Page, added some properties to the button style to cut down redundancy. Added text and image for each button, and added the borders between the buttons.