Lab6 Completed
Added GetForecast method on weather retriver that does the same thing as observations except uses different link and puts result into a ForecastRootObject.
In MainPage.xaml.cs, added the ForecastDayViewModels to the Forecast collection on MainPageViewModel using GetForecast and loops.
In MainPage.xaml, added formatting for the GridView and Image icon to fit and respond to resizing.