Data Binding Slides steps
I was debugging my button page because pressing the button was not increasing the number in the text block on main page, so I knew there was a bug somewhere. I copied all of the data binding steps from Lab 6 except doing the unload and added the Package references. After adding them, the property changed using statement in Data.cs now worked and the data was updated on main page.