Upgrades Formatting
Added Grid instead of relative panel in the DataTemplate because the space was limited in the gridview item. Added a third helper to see if it would make each item stretch all the way to the scroll bar because I want one helper per row, but it is sticking two in each row. Also tried making the second column in the gridview item grid to * so it goes all the way to the scroll bar but it seems like the gridview item is limited to half of the row.