Upgrades Functionality!
Separated the hire/upgrade button from the data binding since there was no way to tell which helper was being upgraded from the button. And added a bunch of new information to each helper along with formatting.

Added functionality to CPS of helper and total CPS. Using timer, cps is added to total commit count each second.

Added a bunch of fields on helper to accomodate the extra information needed in upgrades page data binding.