Created Folders and Added Upgrades Functionality
Folders one for Pages one for classes, moved all files except Main Page into those two folders.

Added functionality and some upgrades to start out with. Lots more helpers and upgrades will be added later but this is the idea. They will be added in the data.cs class in the AddHelpers and AddUpgrades methods.