Formatting + Upgrades Data binding changes
Changed frame on main page to not have margins so the horizontal rectangle below "Helpers" on Upgrades page goes all the way left. Updated margins on several Upgrades page controls to preserve formatting.

For the data binding, I reverted back to an older version where the button was part of the binding, but used a casting technique to figure out which button was being pressed even though they all call the same click method.