Options Page, Formatting, Sounds
Added music track which loops when it finishes. Sound whenever a navigation button, back button, or options is clicked. Added background.

Options page added, with sliders that change the volume of music and click sound.

Formatting on other pages so the title textblock and black line under is consistent across pages.